LayoutTests/media/media-fragments/TC0090-TC0099-expected.txt


Title: TC0090
Fragment: '&&t=3,7'
Comment: Sprinkling & is OK.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
EXPECTED (video.initialTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

Title: TC0091
Fragment: '&t=3,7&'
Comment: Sprinkling & is OK.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
EXPECTED (video.initialTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

Title: TC0092
Fragment: 't%3d10'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.initialTime == '0') OK

Title: TC0093
Fragment: 't=10%26'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.initialTime == '0') OK

Title: TC0094
Fragment: 't=3,7,'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.initialTime == '0') OK

END OF TEST
